Update magic_wifi.ino
Adding resolution when DSC is requested the firmware version. This allows a user that forgets the alt/az resolutions to find out by using a wifi terminal to connect to the DSC and sending "V" to the DSC. The DSC will respond with the firmware version and alt/az resolution.